ops: use uint64 for DocsAndBytes bytes
32 bit integers overflow during moderately large transactions, resulting in incorrect stats being
reported when the number of bytes in a single transaction exceeds the maximum a 32 bit integer can
represent. Using a 64 bit integer allows for much larger byte counts.

+++ crates/runtime/src/derive/combine.rs
@@ -78,7 +78,7 @@ impl Forward {
             }
             let read_stats = &mut self.read_stats[read.transform as usize];
             read_stats.docs_total += 1;
-            read_stats.bytes_total += read.doc_json.len() as u32;
+            read_stats.bytes_total += read.doc_json.len() as u64;
         }
 
         // Build an intent to tell the response loop about a request::Flush.
@@ -365,7 +365,7 @@ impl State {
         memtable.add(doc, false)?;
 
         self.publish_stats.docs_total += 1;
-        self.publish_stats.bytes_total += published.doc_json.len() as u32;
+        self.publish_stats.bytes_total += published.doc_json.len() as u64;
 
         Ok(())
     }
@@ -385,7 +385,7 @@ impl State {
             let doc_json = serde_json::to_string(&doc).expect("document serialization cannot fail");
 
             self.drain_stats.docs_total += 1;
-            self.drain_stats.bytes_total += doc_json.len() as u32;
+            self.drain_stats.bytes_total += doc_json.len() as u64;
 
             let key_packed = crate::extract_packed_node(
                 &doc,
